+++ src/assets/images/tree_animation.gif
