--- src/components/Detector/Detector2.css
@@ -1,9 +1,0 @@
-.crop-container,
-.croper {
-    border: 2px solid;
-    width: 400px;
-    height: 300px;
-}
-.controles {
-    height: 50px;
-}
